[RFR-275] Move Map to shared library
