+++ measuring-client/src/main/java/de/cyface/app/ui/nav/controller/EventDeleteController.java
@@ -36,7 +36,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 
 import de.cyface.app.R;
-import de.cyface.app.ui.Map;
+import de.cyface.app.utils.Map;
 import de.cyface.persistence.DefaultPersistenceBehaviour;
 import de.cyface.persistence.DefaultPersistenceLayer;
 import de.cyface.persistence.content.BaseColumns;
