+++ measuring-client/src/main/res/values/strings.xml
@@ -83,20 +83,18 @@
   <string name="settings_sensor_frequency_unit" translatable="false">Hz</string>
 
   <!-- Measurement page -->
-  <string name="modality_type">Modality type</string>
   <string name="mark_all_menu_item">Select all items</string>
   <string name="delete_marked_menu_item">Delete marked</string>
   <string name="measurement">Measurement</string>
   <string name="toast_export">Exported:</string>
   <string name="title_menu_item_export">Export all</string>
-    <!-- Keep <= 60 characters as Toasts are limited on Android 12+. -->
   <string name="export_data">Exporting data …</string>
   <string name="export_data_no_permission">Access to external memory (SD) required.</string>
   <string name="delete_data">Deleting marked measurements …</string>
   <string name="delete_data_non_selected">No measurements selected. Multiple selection by long press.</string>
   <string name="toast_selection_mode_enabled">Selection mode enabled</string>
   <string name="toast_selection_mode_disabled">Selection mode disabled</string>
-  <string name="toast_no_locations_found">No locations found</string>
+  <!--<string name="toast_no_locations_found">No locations found</string>-->
   <string name="toast_measurement_deletion_success">Copies of the measurements on the server are not affected.</string>
   <string name="toast_item_deletion_success">The selected items were deleted.</string>
   <string name="toast_select_location">Please select a location</string>
