--- ui/r4r/src/main/java/de/cyface/app/r4r/ui/capturing/dialog/CalibrationDialogListener.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017 Cyface GmbH
- *
- * This file is part of the Cyface App for Android.
- *
- * The Cyface App for Android is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * The Cyface App for Android is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with the Cyface App for Android. If not, see <http://www.gnu.org/licenses/>.
- */
-package de.cyface.app.r4r.ui.capturing.dialog
-
-/**
- * An interface for listeners who are interested in the calibration dialog shown when capturing is started.
- *
- * @author Armin Schnabel
- * @version 2.0.0
- * @since 1.0.0
- */
-interface CalibrationDialogListener {
-    fun onCalibrationDialogFinished()
-}
