+++ utils/build.gradle
@@ -23,9 +23,19 @@
  * @version 1.0.0
  * @since 3.2.0
  */
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
+apply plugin: 'com.android.library'
+apply plugin: 'io.sentry.android.gradle' // Exception tracking when using proguard
+apply plugin: 'org.jetbrains.kotlin.android'
+
+buildscript {
+    repositories {
+        mavenCentral()
+    }
+
+    dependencies {
+        // Exception tracking when using proguard
+        classpath "io.sentry:sentry-android-gradle-plugin:$rootProject.ext.sentryAndroidGradlePluginVersion"
+    }
 }
 
 android {
@@ -45,9 +55,6 @@ android {
         //buildConfigField "String", "GitHash", '"${getGitHash()}"'
         // we don't describe this provider build variable in the README so it's probably deprecated
         // buildConfigField("String", "provider", '"de.cyface.app.provider"')
-
-        // If our terms change that much that they to be re-accepted, increase this (see Confluence!)
-        //buildConfigField "int", "currentTerms", "4"
 
         // Ensure that the backend's "CyfaceFull" flavor is always used
         missingDimensionStrategy 'project', 'cyface'
@@ -78,12 +85,38 @@ android {
     }
 }
 
+// Exception tracking
+// FIXME: we should probably add something to see that these crashes come from the RFR app!
+sentry {
+    // Enables or disables the automatic upload of mapping files
+    // during a build.  If you disable this, you'll need to manually
+    // upload the mapping files with sentry-cli when you do a release.
+    autoUpload = true
+
+    // Disables or enables the automatic configuration of Native Symbols
+    // for Sentry. This executes sentry-cli automatically so
+    // you don't need to do it manually.
+    // Default is disabled.
+    uploadNativeSymbols = false
+
+    // Does or doesn't include the source code of native code for Sentry.
+    // This executes sentry-cli with the --include-sources param. automatically so
+    // you don't need to do it manually.
+    // Default is disabled.
+    includeNativeSources = false
+}
+
 dependencies {
+    // Exception tracking
+    implementation "io.sentry:sentry-android:$rootProject.ext.sentryAndroidVersion"
 
     // To change the material design
     //implementation "androidx.appcompat:appcompat:$rootProject.ext.androidxAppCompatVersion"
     // To use material elements (tabs, nav bar, slider, etc.)
     implementation "com.google.android.material:material:$rootProject.ext.materialVersion"
+    // To use Google Map
+    implementation "com.google.android.gms:play-services-maps:$rootProject.ext.mapPlayServicesVersion"
+    implementation "com.google.android.gms:play-services-location:$rootProject.ext.locationPlayServicesVersion"
     // Generated by the RFR-UI
     //implementation 'androidx.core:core-ktx:1.7.0'
 
@@ -91,7 +124,7 @@ dependencies {
     implementation "de.cyface:android-utils:$rootProject.ext.cyfaceUtilsVersion"
     implementation project(':datacapturing')
     //implementation project(':synchronization')
-    //implementation project(':persistence')
+    implementation project(':persistence')
     //implementation project(':energy_settings')
 
     // Dependencies for instrumentation tests
